whisper_base/requirements.txt
sagemaker>=2.180.0,<3.0.0
boto3>=1.28.0
openai-whisper
torch
soundfile
scipy
python-dotenv